+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - General
   - Bump clj-kondo to `2023.12.16-20240202.114914-15`.
 
+- Editor
+  - Fix documentation resolve not working for clients without hover markdown support.
+
 ## 2024.02.01-11.01.59
 
 - General